+++ packages/esm-patient-queues-app/src/active-visits/active-visits-print/active-visits-print.component.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
-import { MappedPatientQueueEntry } from '../patient-queues.resource';
 import styles from './active-visits-print.scss';
 import { QRCodeSVG } from 'qrcode.react';
 
 import PatientQueueDetailsTable from './patient-queue-details-table.component';
+import { PatientQueue } from '../../types/patient-queues';
 
 interface VisitCardToPrintProps {
-  queueEntry: MappedPatientQueueEntry;
+  queueEntry: PatientQueue;
 }
 
 export function VisitCardToPrint({ queueEntry }: VisitCardToPrintProps) {
-  const patientUIC = queueEntry.identifiers.find((item) => item.display.includes('Patient Unique  ID Code'));
+  const patientUIC = queueEntry.patient?.identifiers.find((item) => item.display.includes('Patient Unique  ID Code'));
 
   return (
     <div className={styles.printPage}>
@@ -18,7 +18,7 @@ export function VisitCardToPrint({ queueEntry }: VisitCardToPrintProps) {
         <h3 style={{ paddingBottom: '8px' }}>Visit Registration Receipt</h3>
         <PatientQueueDetailsTable queueEntry={queueEntry} />
         <div style={{ margin: '25px' }} className={styles.name}>
-          {queueEntry.identifiers.length > 0 && patientUIC ? (
+          {queueEntry?.patient?.identifiers.length > 0 && patientUIC ? (
             <QRCodeSVG value={patientUIC?.display.split('=')[1]} />
           ) : (
             <> NA</>
