+++ packages/esm-patient-queues-app/src/active-visits/notes-action-menu.components.tsx
@@ -3,7 +3,6 @@ import { Button } from '@carbon/react';
 import { CatalogPublish } from '@carbon/react/icons';
 import { useTranslation } from 'react-i18next';
 import { showModal } from '@openmrs/esm-framework';
-import { MappedPatientQueueEntry } from './patient-queues.resource';
 import { PatientQueue } from '../types/patient-queues';
 
 interface NotesActionsMenuProps {
