+++ packages/esm-patient-queues-app/src/active-visits/notes-dialog.component.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { Button, Form, ModalBody, ModalFooter, ModalHeader } from '@carbon/react';
 import { useTranslation } from 'react-i18next';
 import styles from './change-status-dialog.scss';
-import { MappedPatientQueueEntry } from './patient-queues.resource';
+import { PatientQueue } from '../types/patient-queues';
 
 interface NotesDialogProps {
-  queueEntry: MappedPatientQueueEntry;
+  queueEntry: PatientQueue;
   closeModal: () => void;
 }
