+++ packages/esm-patient-queues-app/src/helpers/functions.ts
@@ -1,16 +1,5 @@
 import { OpenmrsResource } from '@openmrs/esm-framework';
-import { MappedQueuePriority } from '../active-visits/active-visits-table.resource';
 
-export const getTagType = (priority: string) => {
-  switch (priority as MappedQueuePriority) {
-    case 'Emergency':
-      return 'red';
-    case 'Not Urgent':
-      return 'green';
-    default:
-      return 'gray';
-  }
-};
 
 export const buildStatusString = (status: string) => {
   if (!status) {
@@ -70,18 +59,6 @@ export const getGender = (gender, t) => {
       return gender;
   }
 };
-
-export function findObsByConceptUUID(arr: Array<OpenmrsResource>, ids: Array<string>) {
-  for (const visit of arr) {
-    return visit.obs.filter((o) => {
-      return ids.includes(o.concept.uuid);
-    });
-  }
-}
-
-export function timeDiffInMinutes(date1: Date, date2: Date) {
-  return Math.round((date1.getTime() - date2.getTime()) / (1000 * 3600 * 24));
-}
 
 export const getProviderTagColor = (entryProvider: string, loggedInProviderName: string) => {
   if (entryProvider === loggedInProviderName) {
