+++ packages/esm-patient-queues-app/src/helpers/helpers.tsx
@@ -1,17 +1,8 @@
 import { getGlobalStore } from '@openmrs/esm-framework';
 import { useEffect, useState } from 'react';
-import { AppointmentSummary } from '../types';
 import { PatientQueue } from '../types/patient-queues';
 
-export const getServiceCountByAppointmentType = (
-  appointmentSummary: Array<AppointmentSummary>,
-  appointmentType: string,
-) => {
-  return appointmentSummary
-    .map((el) => Object.entries(el.appointmentCountMap).flatMap((el) => el[1][appointmentType]))
-    .flat(1)
-    .reduce((count, val) => count + val, 0);
-};
+
 
 const initialServiceNameState = { serviceName: localStorage.getItem('queueServiceName') };
 const initialServiceUuidState = { serviceUuid: localStorage.getItem('queueServiceUuid') };
